training/openai_agents_api/pyproject.toml
[project]
name = "openai-agents-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "graphviz>=0.20.3",
    "nest-asyncio>=1.6.0",
    "openai-agents>=0.0.7",
    "python-dotenv>=1.1.0",
]
